+++ changelog.md
@@ -14,6 +14,7 @@ and this project adheres to([https://semver.org/spec/v2.0.0.html](https://semver
 ### Changed
 - Incremented version in `package.json`, `package-lock.json`.
 - Hidden `Annotation Panel` control.
+- Hidden exports of `PPTist, pptx, image, json` to keep only PDF export.
 
 ### Fixed
 -